+++ src/main/java/fuzs/goldenagecombat/handler/SweepAttackHandler.java
@@ -1,5 +1,6 @@
 package fuzs.goldenagecombat.handler;
 
+import fuzs.goldenagecombat.GoldenAgeCombat;
 import fuzs.goldenagecombat.registry.ModRegistry;
 import net.minecraft.core.particles.ParticleTypes;
 import net.minecraft.server.level.ServerLevel;
@@ -14,25 +15,26 @@ import net.minecraft.world.entity.decoration.ArmorStand;
 import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.item.enchantment.EnchantmentHelper;
 import net.minecraft.world.phys.AABB;
+import net.minecraftforge.common.ToolActions;
 import org.jetbrains.annotations.Nullable;
 
 import java.util.List;
 
 public class SweepAttackHandler {
     public static void attackAir(Player player) {
-//        if (!player.isAttackAvailable(1.0f)) {
-//            return;
-//        }
-//        player.swing(InteractionHand.MAIN_HAND);
-//        float attackStrength = (float) player.getAttribute(Attributes.ATTACK_DAMAGE).getValue();
-//        if (attackStrength > 0.0f && player.checkSweepAttack()) {
-//            float attackReach = (float) getCurrentAttackReach(player);
-//            double moveX = (double)(-Mth.sin(player.getYRot() * ((float)Math.PI / 180))) * 2.0;
-//            double moveZ = (double)Mth.cos(player.getYRot() * ((float)Math.PI / 180)) * 2.0;
-//            AABB aABB = player.getBoundingBox().inflate(1.0, 0.25, 1.0).move(moveX, 0.0, moveZ);
-//            sweepAttack(player, aABB, attackReach, attackStrength, null);
-//        }
-//        player.resetAttackStrengthTicker();
+        double d0 = player.walkDist - player.walkDistO;
+        if (!player.isOnGround() || !(d0 < player.getSpeed())) return;
+        float attackDamage = (float) player.getAttribute(Attributes.ATTACK_DAMAGE).getValue();
+        ClassicCombatHandler.disableCooldownPeriod(player);
+        if (attackDamage > 0.0f && player.getAttackStrengthScale(0.5F) > 0.9F && checkSweepAttack(player)) {
+            float attackReach = (float) getCurrentAttackReach(player);
+            double moveX = (double)(-Mth.sin(player.getYRot() * ((float)Math.PI / 180))) * 2.0;
+            double moveZ = (double)Mth.cos(player.getYRot() * ((float)Math.PI / 180)) * 2.0;
+            AABB aABB = player.getItemInHand(InteractionHand.MAIN_HAND).getSweepHitBox(player, player).move(moveX, 0.0, moveZ);
+            sweepAttack(player, aABB, attackReach, attackDamage, null);
+        }
+        // also resets attack ticker
+        player.swing(InteractionHand.MAIN_HAND);
     }
 
     public static void sweepAttack(Player player, AABB aABB, float currentAttackReach, float baseAttackDamage, @Nullable Entity target) {
@@ -53,14 +55,25 @@ public class SweepAttackHandler {
         }
     }
 
+    public static boolean checkSweepAttack(Player player) {
+        if (GoldenAgeCombat.CONFIG.server().combatTests.noSneakSweeping && player.isShiftKeyDown()) {
+            return false;
+        }
+        if (GoldenAgeCombat.CONFIG.server().combatTests.sweepingRequired) {
+            return EnchantmentHelper.getSweepingDamageRatio(player) > 0.0f;
+        }
+        return player.getItemInHand(InteractionHand.MAIN_HAND).canPerformAction(ToolActions.SWORD_SWEEP);
+    }
+
     public static double getCurrentAttackReach(Player player) {
+        if (!GoldenAgeCombat.CONFIG.server().attributes.attackReach) return 3.0;
         double attackReach = player.getAttribute(ModRegistry.ATTACK_REACH_ATTRIBUTE.get()).getValue();
-//        if (this.minecraft.gameMode.hasFarPickRange()) {
-//            attackReach += 0.5;
-//        }
-//        if (player.isCrouching()) {
-//            attackReach -= 0.5;
-//        }
+        if (player.isCreative()) {
+            attackReach += 0.5;
+        }
+        if (player.isCrouching()) {
+            attackReach -= 0.5;
+        }
         return attackReach;
     }
 }