+++ src/main/java/fuzs/goldenagecombat/mixin/client/GameRendererMixin.java
@@ -5,6 +5,7 @@ import fuzs.goldenagecombat.registry.ModRegistry;
 import net.minecraft.client.Minecraft;
 import net.minecraft.client.renderer.GameRenderer;
 import net.minecraft.world.entity.Entity;
+import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.entity.projectile.ProjectileUtil;
 import net.minecraft.world.level.ClipContext;
 import net.minecraft.world.phys.AABB;
@@ -29,22 +30,13 @@ public abstract class GameRendererMixin {
 
     @Inject(method = "pick", at = @At("HEAD"), cancellable = true)
     public void pick$head(float partialTicks, CallbackInfo callbackInfo) {
-        if (!GoldenAgeCombat.CONFIG.server().adjustments.swingThroughGrass && !GoldenAgeCombat.CONFIG.server().attributes.increasedAttackReach) return;
+        if (!GoldenAgeCombat.CONFIG.server().combatTests.swingThroughGrass && !GoldenAgeCombat.CONFIG.server().attributes.attackReach) return;
         Entity entity = this.minecraft.getCameraEntity();
         if (entity != null && this.minecraft.level != null) {
             this.minecraft.getProfiler().push("pick");
             this.minecraft.crosshairPickEntity = null;
             double maxPickRange = this.minecraft.gameMode.getPickRange();
-            double maxEntityPickRange = 3.0;
-            if (GoldenAgeCombat.CONFIG.server().attributes.increasedAttackReach) {
-                maxEntityPickRange = this.minecraft.player.getAttribute(ModRegistry.ATTACK_REACH_ATTRIBUTE.get()).getValue();
-                if (this.minecraft.gameMode.hasFarPickRange()) {
-                    maxEntityPickRange += 0.5;
-                }
-                if (entity.isCrouching()) {
-                    maxEntityPickRange -= 0.5;
-                }
-            }
+            double maxEntityPickRange = entity == this.minecraft.player && GoldenAgeCombat.CONFIG.server().attributes.attackReach ? this.getCurrentAttackReach(this.minecraft.player) : 3.0;
             Vec3 viewVector = entity.getViewVector(1.0F);
             Vec3 eyePosition = entity.getEyePosition(partialTicks);
             Vec3 pickVector = eyePosition.add(viewVector.x * maxEntityPickRange, viewVector.y * maxEntityPickRange, viewVector.z * maxEntityPickRange);
@@ -63,7 +55,7 @@ public abstract class GameRendererMixin {
             HitResult outlineHitResult = entity.pick(maxPickRange, partialTicks, false);
             // when trying to pick an entity, pick a second time with collider context, so we are able to avoid e.g. tall grass
             // doing this the other way around does not work, as e.g. fences will be picked from their collision box when an entity is standing behind them
-            if (GoldenAgeCombat.CONFIG.server().adjustments.swingThroughGrass && entityhitresult != null && outlineHitResult.getType() != HitResult.Type.MISS) {
+            if (GoldenAgeCombat.CONFIG.server().combatTests.swingThroughGrass && entityhitresult != null && outlineHitResult.getType() != HitResult.Type.MISS) {
                 HitResult colliderHitResult = pick(entity, maxPickRange, partialTicks);
                 if (colliderHitResult.getType() == HitResult.Type.MISS) {
                     outlineHitResult = colliderHitResult;
@@ -82,6 +74,17 @@ public abstract class GameRendererMixin {
             this.minecraft.getProfiler().pop();
         }
         callbackInfo.cancel();
+    }
+
+    private double getCurrentAttackReach(Player player) {
+        double attackReach = player.getAttribute(ModRegistry.ATTACK_REACH_ATTRIBUTE.get()).getValue();
+        if (this.minecraft.gameMode.hasFarPickRange()) {
+            attackReach += 0.5;
+        }
+        if (player.isCrouching()) {
+            attackReach -= 0.5;
+        }
+        return attackReach;
     }
 
     private static double getEntityPickRange(EntityHitResult entityhitresult, Vec3 eyePosition, Vec3 pickVector) {