+++ src/main/java/fuzs/goldenagecombat/mixin/client/accessor/MultiPlayerGameModeAccessor.java
@@ -1,0 +1,11 @@
+package fuzs.goldenagecombat.mixin.client.accessor;
+
+import net.minecraft.client.multiplayer.MultiPlayerGameMode;
+import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.gen.Invoker;
+
+@Mixin(MultiPlayerGameMode.class)
+public interface MultiPlayerGameModeAccessor {
+    @Invoker
+    void callEnsureHasSentCarriedItem();
+}